+++ src/types/index.ts
@@ -1,7 +1,10 @@
 export * from './vote';
 export * from './metadata/election';
 export * from './metadata/account';
-export * from './election';
+// Beware the correct order for avoiding circular dependency
+export * from './election/unpublished';
+export * from './election/election';
+export * from './election/published';
 export * from './account';
 export * from './census/plain';
 export * from './census/published';