+++ CHANGELOG.md
@@ -18,7 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use voting endpoint instead of generic submitTx.
-- Naming for client initialization options changed
+- Naming for client initialization options changed.
+- There are now the new `UnpublishedElection` and `PublishedElection` classes
+which extend from the abstract `Election` class.
 
 ## [0.0.1-alpha] - 2022-12-01